Add _Py_IsValidFD() helper function
